requirements.txt
gevent
influxdb
PyYAML
